[azureCI] Add a coverage job
This commit uses the pre-existing templates and build/test
scripts to include another weekly CI job that runs on azure.

In the future, it may be ideal to switch all build and test runs
to using the rccl-test-infra framework. For now, we just focus
on coverage environment variable sweeps.

This commit also publishes the results for easy access and downloads
on the latest RCCL commit, once per week.